requirements.txt
streamlit
plotly
pyserial
smtplib
email
datetime
threading
queue
csv
os
time
